refactor(plugins/github): create package directory
