+++ plugins/github/pkg/github/secrets.go
@@ -1,4 +1,4 @@
-package main
+package github
 
 import (
 	"regexp"
