+++ plugins/github/pkg/github/server.go
@@ -1,4 +1,4 @@
-package main
+package github
 
 import (
 	"bufio"
